upload file zip has encrypt second
